backend/requirements.txt
# Web 框架
flask>=2.3.3
flask-cors>=4.0.0
Werkzeug>=2.3.7
gunicorn>=20.1.0  # GAE需要gunicorn作为WSGI服务器

# 深度学习框架
tensorflow-cpu>=2.16.0  # 使用CPU版本减小包大小
torch>=2.2.0  # 代码中使用了torch，不能注释掉
transformers==4.30.2
openai-whisper>=20231117
fer>=22.5.0

# 数据处理
numpy>=1.24.3
pandas>=2.0.0  # 数据处理常用库

# 工具依赖
packaging>=20.0
setuptools>=65.5.1
opencc-python-reimplemented>=0.1.7  # 用于简繁中文转换
python-magic>=0.4.27  # 用于文件类型检测
logging>=0.5.1.2  # 日志处理

# 视频和音频处理
opencv-python-headless>=4.8.0  # 替换为无头版本
moviepy>=1.0.3
Pillow>=10.0.0

# 安全相关
werkzeug>=2.3.7  # 安全文件名处理

# 文件处理
tempfile  # 标准库，不需要安装
base64  # 标准库，不需要安装
